Put all deck related functions inside of deck module.
Optimise txscapy function to avoid calling a function on every import.
