+++ ooni/deck/__init__.py
@@ -1,0 +1,1 @@
+from .deck import NGDeck
