+++ ooni/utils/txscapy.py
@@ -4,6 +4,7 @@ import random
 from twisted.internet import fdesc
 from twisted.internet import reactor
 from twisted.internet import defer, abstract
+
 from scapy.config import conf
 from scapy.all import RandShort, IP, IPerror, ICMP, ICMPerror, TCP, TCPerror, UDP, UDPerror
 
@@ -15,58 +16,45 @@ from ooni.utils.net import getDefaultIface, getAddresses
 from ooni.settings import config
 
 
-def pcapdnet_installed():
-    """
-    Checks to see if libdnet or libpcap are installed and set the according
-    variables.
+# Check to see if libdnet or libpcap are installed and set the according
+# variables.
 
-    Returns:
+# In debian libdnet is called dumbnet instead of dnet, but scapy is
+# expecting "dnet" so we try and import it under such name.
+try:
+    import dumbnet
 
-        True
-            if pypcap and libdnet are installed
+    sys.modules['dnet'] = dumbnet
+except ImportError:
+    pass
 
-        False
-            if one of the two is absent
-    """
-    # In debian libdnet is called dumbnet instead of dnet, but scapy is
-    # expecting "dnet" so we try and import it under such name.
-    try:
-        import dumbnet
+try:
+    conf.use_pcap = True
+    conf.use_dnet = True
+    from scapy.arch import pcapdnet
 
-        sys.modules['dnet'] = dumbnet
-    except ImportError:
-        pass
+    config.pcap_dnet = True
 
-    try:
-        conf.use_pcap = True
-        conf.use_dnet = True
-        from scapy.arch import pcapdnet
+except ImportError as e:
+    log.err(e.message + ". Pypcap or dnet are not properly installed. Certain tests may not work.")
+    config.pcap_dnet = False
+    conf.use_pcap = False
+    conf.use_dnet = False
 
-        config.pcap_dnet = True
+# This is required for unix systems that are different than linux (OSX for
+# example) since scapy explicitly wants pcap and libdnet installed for it
+# to work.
+try:
+    from scapy.arch import pcapdnet
+except ImportError:
+    log.err("Your platform requires having libdnet and libpcap installed.")
+    raise LibraryNotInstalledError
 
-    except ImportError as e:
-        log.err(e.message + ". Pypcap or dnet are not properly installed. Certain tests may not work.")
-        config.pcap_dnet = False
-        conf.use_pcap = False
-        conf.use_dnet = False
+_PCAP_DNET_INSTALLED = config.pcap_dnet
 
-    # This is required for unix systems that are different than linux (OSX for
-    # example) since scapy explicitly wants pcap and libdnet installed for it
-    # to work.
-    try:
-        from scapy.arch import pcapdnet
-    except ImportError:
-        log.err("Your platform requires having libdnet and libpcap installed.")
-        raise LibraryNotInstalledError
-
-    return config.pcap_dnet
-
-
-if pcapdnet_installed():
+if _PCAP_DNET_INSTALLED:
     from scapy.all import PcapWriter
-
 else:
-
     class DummyPcapWriter:
         def __init__(self, pcap_filename, *arg, **kw):
             log.err("Initializing DummyPcapWriter. We will not actually write to a pcapfile")
@@ -78,7 +66,6 @@ else:
     PcapWriter = DummyPcapWriter
 
 from scapy.all import Gen, SetGen, MTU
-
 
 class ScapyFactory(abstract.FileDescriptor):
     """
